fixed func declaration

+++ matrices.c
@@ -76,7 +76,7 @@ int testMatrix(int** matrix_1, int** matrix_2) {
     return failures;
 }
 
-int** multiplyMatrix(int **matrix_1, int **matrix_2) {
+int** multiplyMatrix(uint8_t **matrix_1, uint8_t **matrix_2) {
     printf("\t\tAttempting to multiply matrices...\n");
     int** result = calloc(DEFAULT_SIZE, sizeof(int*));
     for (int i = 0; i < DEFAULT_SIZE; i++) {
